--- a/Gemfile
+++ b/Gemfile
@@ -53,6 +53,10 @@ group :test do
   gem "webmock"
 end
 
+group :test, :development, :staging do
+  gem "faker", require: false
+end
+
 group :staging, :production do
   gem "rack-timeout"
 end