Adds some basic seeds data from Faker. I'm not sure if putting faker in the general gemset is the best approach? I figured we may need to run this on heroku, so I just put `require: false` to prevent it from getting into the rails app by mistake. Any other ideas?