@evan-007 2015-11-07
@samnang @ajkamel 

It's a little faster when wrapping the seeds in a transaction:

before:

```
bundle exec rake db:seed  2.48s user 0.56s system 98% cpu 3.078 total
```

after:
```
bundle exec rake db:seed  2.26s user 0.49s system 98% cpu 2.783 total
```

nice insight Samnang! 

okay to merge?